feat: add event modal
